Refactor v2 eval value helpers
